Correctly detect mismatches in the number of copied bytes

+++ src/XrdClient/Xrdcp.cc
@@ -685,7 +685,7 @@ int doCp_xrd2xrd(XrdClient **xrddest, const char *src, const char *dst) {
       cout << endl;
    }
 
-   if ((unsigned)cpnfo.len != bytesread) retvalue = 13;
+   if (cpnfo.len != bytesread) retvalue = 13;
 
 #ifdef HAVE_XRDCRYPTO
    if (md5) MD_5->Final();
@@ -957,7 +957,7 @@ int doCp_xrd2loc(const char *src, const char *dst) {
       cout << endl;
    }
 
-   if ((unsigned)cpnfo.len != bytesread) retvalue = 13;
+   if (cpnfo.len != bytesread) retvalue = 13;
 
 #ifdef HAVE_XRDCRYPTO
    if (md5) MD_5->Final();
